Add PropTypes and use redux as share state
